requirements.txt
Authlib==0.10
Click==7.0
Flask-Caching==1.4.0
Flask-Cors==3.0.6
Flask-JWT-Extended==3.10.0
Flask-SQLAlchemy==2.3.2
Flask==1.0.2
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.6.4
PyYAML==3.12
SQLAlchemy==1.2.8
Werkzeug==0.14.1
argcomplete==1.9.3
asn1crypto==0.24.0
blinker==1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.5
facebook-sdk==3.0.0
flask-marshmallow==0.9.0
flask-swagger==0.2.13
graypy==0.3.1
idna==2.7
itsdangerous==1.1.0
marshmallow-sqlalchemy==0.14.0
marshmallow==2.15.3
mysql-connector-python==8.0.11
protobuf==3.6.1
pycparser==2.19
python-dateutil==2.6.1
redis==2.10.6
requests==2.19.1
sentry-sdk==0.6.9
six==1.12.0
urllib3==1.23
tabulate==0.8.3
